+++ ProyectoFinal/app/Http/Controllers/CategoriaControler.php
@@ -48,6 +48,10 @@ class CategoriaControler extends Controller
      */
     public function show($id){
         try {
+            $checkCat = Categoria::find($id);
+            if($checkCat==null){
+                return response()->json(['error' => 'La ID categoria no existe'], 404);
+            }
             $tupla = Categoria::findOrFail($id);
                 return response()->json(['status' => 'success', 'result' => $tupla], 200);
             }catch (\Exception $e){
@@ -110,6 +114,10 @@ class CategoriaControler extends Controller
      */
     public function borra($id){
         try {
+            $checkCat = Categoria::find($id);
+            if($checkCat==null){
+                return response()->json(['error' => 'La ID categoria no existe'], 404);
+            }
             $tupla = Categoria::findOrFail($id)->delete();
             return response()->json(['status' => 'success', 'result' => $tupla], 200);
         }catch (\Exception $e){
@@ -235,6 +243,10 @@ class CategoriaControler extends Controller
             'filled'=>':attribute no pot estar buit',
             'unique'=>'Camp :attribute amb valor :input ja hi es'
         ];
+        $checkCat = Categoria::find($id);
+        if($checkCat==null){
+            return response()->json(['error' => 'La ID categoria no existe'], 404);
+        }
         $validacio=Validator::make($request->all(),$reglesvalidacio,$missatges);
         if(!$validacio->fails()){
             $tupla->update($request->all());
